+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.1.37
+- Move Ats `entries.dart` file to `entry.dart` and create a new folder `entries` to put inside all Models of Ats Entries
+- Create `AtsVolume` model
+- Add `volumeHistory` to `AtsEntry
+
+## 2.1.36
+
+- Added `modbusParameter` type `ModbusParameter` model in `DeviceCommand` model
+- Added `modbusPort` in `DeviceCommand` model
+
 ## 2.1.35
 - Add `fuelSubtype` - `density` - `temperature` - tanksImages in `AtsReceptionProduct`